+++ package.json
@@ -14,8 +14,13 @@
     "http-server": "^0.11.1",
     "yargs": "^12.0.2"
   },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^6.1.4"
+  },
   "scripts": {
     "start": "hugo server -D",
+    "test": "mocha scripts/lib/test",
     "build:local": "npm run build:graphs && hugo",
     "build:prod": "npm run build:graphs && hugo --gc --minify",
     "build:preview": "npm run build:graphs && hugo --gc --minify -b $DEPLOY_PRIME_URL",
